Focus first on dependence on switching time
Factor out script that looks at bidirectional estimates...
